requirements.txt
flask
flask-cors
numpy
dotenv-python